poc/modbus/requirements-bridge.txt
# Modbus + Kafka in OpenShift (Python 3.12)
pymodbus==3.6.8
kafka-python-ng==2.2.3
